+++ framework/widgets/Breadcrumbs.php
@@ -147,7 +147,7 @@ class Breadcrumbs extends Widget
             }
             $links[] = $this->renderItem($link, isset($link['url']) ? $this->itemTemplate : $this->activeItemTemplate);
         }
-        echo Html::tag($this->tag, implode('', $links), $this->options);
+        return Html::tag($this->tag, implode('', $links), $this->options);
     }
 
     /**